api/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
influxdb-client==1.45.0